Created Segment Tracking plugin.
This plugin is based on the previous plugin we had, but is greatly simplified and is restricted to only track events where the user is identified. It could be extended for other scenarios in the future though.

Tracking via this plugin should only be done when client-side tracking is not possible (e.g. when the server redirects instead of rendering).